# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Booking Clone API",
    "description": "REST API za aplikaciju za rezervaciju apartmana.\n\nTri uloge: **USER** rezerviše, **HOST** izdaje i vodi svoje apartmane,\n**ADMIN** upravlja šifarnicima.\n\n**Autentifikacija.** Prijava vraća *access token* koji traje 15 minuta i putuje\nu zaglavlju `Authorization: Bearer <token>`. Uz njega ide i *refresh token*, ali\non stoji u `httponly` kolačiću i ne čita se iz JavaScript-a; `POST /auth/refresh`\nga troši i izdaje nov par, pa jedan refresh token vredi tačno jednom.\n\nU Swagger UI-ju se prijavljujete dugmetom **Authorize**, unosom email adrese i\nlozinke.\n\n**Greške.** Svaka greška ima isti oblik: `detail` je rečenica na engleskom,\n`code` je stabilna oznaka koju frontend prevodi na jezik korisnika, a `params`\nnosi vrednosti koje ulaze u prevedenu rečenicu. Zahvaljujući tome poruka o\ngrešci nikad nije sklopljena od engleskih parčića unutar srpske rečenice.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/apartments": {
      "get": {
        "tags": [
          "apartments"
        ],
        "summary": "Pretraga apartmana",
        "description": "Javna, paginirana lista aktivnih apartmana.\n\nFilteri se slažu jedan na drugi. Ako se pošalju oba datuma, u rezultatu\nostaju samo apartmani slobodni za ceo taj raspon. Obrisani apartmani se ne\nprikazuju nigde.",
        "operationId": "get_apartments_apartments_get",
        "parameters": [
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page Number"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "address",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "title": "Address"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "name": "price_per_night_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Price Per Night Min"
            }
          },
          {
            "name": "price_per_night_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Price Per Night Max"
            }
          },
          {
            "name": "max_guests",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Guests"
            }
          },
          {
            "name": "rating_average_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating Average Min"
            }
          },
          {
            "name": "rating_average_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating Average Max"
            }
          },
          {
            "name": "check_in",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Check In"
            }
          },
          {
            "name": "check_out",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Check Out"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasePagedResponse_ApartmentDto_"
                }
              }
            }
          },
          "400": {
            "description": "Zahtev je neispravan, npr. datum odjave pre datuma prijave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apartments"
        ],
        "summary": "Kreiranje apartmana",
        "description": "Pravi apartman i odmah ga postavlja kao aktivan.\n\nAdresa se pri upisu geokodira preko OpenStreetMap Nominatim servisa.\nAko geokodiranje ne uspe, apartman se svejedno pravi, samo bez\nkoordinata: nedostupan spoljni servis ne sme da obori kreiranje.\n\nNovi apartman nema slike, pa je lista `photos` prazna. Slike se\ndodaju posebno, preko `POST /apartments/{id}/photos`.",
        "operationId": "create_apartment_apartments_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApartmentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Zahtev je neispravan, npr. datum odjave pre datuma prijave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Niste prijavljeni ili je token istekao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prijavljeni ste, ali nemate pravo na ovu radnju.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/apartments/{apartment_id}": {
      "patch": {
        "tags": [
          "apartments"
        ],
        "summary": "Izmena apartmana",
        "description": "Menja apartman. Šalje se samo ono što se menja.\n\nPromena adrese, grada ili države pokreće ponovno geokodiranje.\nAko se pošalje `tag_ids`, oznake se zamenjuju u celini, a ne dodaju.",
        "operationId": "update_apartment_apartments__apartment_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Apartment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApartmentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentByIdDto"
                }
              }
            }
          },
          "400": {
            "description": "Zahtev je neispravan, npr. datum odjave pre datuma prijave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Niste prijavljeni ili je token istekao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prijavljeni ste, ali nemate pravo na ovu radnju.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Traženi resurs ne postoji.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "apartments"
        ],
        "summary": "Jedan apartman sa oznakama i slikama",
        "description": "Jedan apartman sa svojim oznakama i slikama.\n\nObrisan apartman se ne vraća: za njega više ne postoji stranica.",
        "operationId": "get_apartment_by_id_apartments__apartment_id__get",
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentByIdDto"
                }
              }
            }
          },
          "404": {
            "description": "Traženi resurs ne postoji.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "apartments"
        ],
        "summary": "Meko brisanje apartmana",
        "description": "Soft delete.\n\nThe row and its photos stay in the database because reservations point at\nthem: a guest still has to see where they stayed and the host still has to\nsee what was booked. Setting deleted_at takes the apartment out of every\nlist and makes it impossible to book again, which is all a delete has to do\nhere.",
        "operationId": "delete_apartment_apartments__apartment_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Niste prijavljeni ili je token istekao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prijavljeni ste, ali nemate pravo na ovu radnju.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Traženi resurs ne postoji.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/apartments/{apartment_id}/photos": {
      "get": {
        "tags": [
          "apartments_photo"
        ],
        "summary": "Sve slike jednog apartmana",
        "description": "Vraća slike apartmana. Javno, jer se iste slike vide i na stranici apartmana.",
        "operationId": "get_apartment_main_photo_apartments__apartment_id__photos_get",
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentPhotoItemDto"
                  },
                  "title": "Response Get Apartment Main Photo Apartments  Apartment Id  Photos Get"
                }
              }
            }
          },
          "404": {
            "description": "Traženi resurs ne postoji.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apartments_photo"
        ],
        "summary": "Otpremanje slika apartmana",
        "description": "Otprema jednu ili više slika za apartman.\n\nSvi fajlovi se prvo provere pa tek onda upisuju, da otpremanje ne\nostavi pola slika na disku ako je poslednja pogrešnog tipa.",
        "operationId": "upload_apartment_photos_apartments__apartment_id__photos_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Apartment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_apartment_photos_apartments__apartment_id__photos_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentPhotoItemDto"
                  },
                  "title": "Response Upload Apartment Photos Apartments  Apartment Id  Photos Post"
                }
              }
            }
          },
          "400": {
            "description": "Zahtev je neispravan, npr. datum odjave pre datuma prijave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Niste prijavljeni ili je token istekao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prijavljeni ste, ali nemate pravo na ovu radnju.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Traženi resurs ne postoji.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "apartments_photo"
        ],
        "summary": "Brisanje slika apartmana",
        "description": "Briše slike po identifikatorima, i iz baze i sa diska.\n\nProlaze samo slike koje stvarno pripadaju tom apartmanu, pa tuđi\nidentifikator ne briše ništa.",
        "operationId": "delete_apartment_photos_apartments__apartment_id__photos_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Apartment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteApartmentPhotosRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Niste prijavljeni ili je token istekao.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Prijavljeni ste, ali nemate pravo na ovu radnju.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Traženi resurs ne postoji.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApartmentCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "title": "Description"
          },
          "address": {
            "type": "string",
            "maxLength": 255,
            "title": "Address"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "title": "City"
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "title": "Country"
          },
          "price_per_night": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Price Per Night"
          },
          "max_guests": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Guests"
          },
          "tag_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Tag Ids"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "address",
          "city",
          "country",
          "price_per_night",
          "max_guests"
        ],
        "title": "ApartmentCreateRequest"
      },
...
      "ApartmentPhotoItemDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "is_main": {
            "type": "boolean",
            "title": "Is Main"
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "is_main"
        ],
        "title": "ApartmentPhotoItemDto",
        "description": "One stored photo. `path` is a url under /static, not a disk path."
      },
...
      "BasePagedResponse_ApartmentDto_": {
        "properties": {
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApartmentDto"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "page_number",
          "page_size",
          "total",
          "items"
        ],
        "title": "BasePagedResponse[ApartmentDto]"
      },
...
      "Body_upload_apartment_photos_apartments__apartment_id__photos_post": {
        "properties": {
          "photos": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Photos"
          }
        },
        "type": "object",
        "required": [
          "photos"
        ],
        "title": "Body_upload_apartment_photos_apartments__apartment_id__photos_post"
      },
      "DeleteApartmentPhotosRequest": {
        "properties": {
          "apartment_photo_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Apartment Photo Ids"
          }
        },
        "type": "object",
        "required": [
          "apartment_photo_ids"
        ],
        "title": "DeleteApartmentPhotosRequest"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Rečenica na engleskom, čitljiva i bez prevoda."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stabilna oznaka greške koju frontend prevodi."
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Vrednosti koje ulaze u prevedenu poruku, npr. `max_guests`."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "The one shape every failed request comes back in.\n\nBuilt by `ExceptionHandlers`, which is the only place that knows what an\nerror looks like on the wire. Declaring it here as well is what puts it in\nthe specification, where it is otherwise invisible.",
        "example": {
          "code": "too_many_guests",
          "detail": "This apartment allows at most 4 guests",
          "params": {
            "max_guests": 4
          }
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  },
...
}